+++ Layers/FF.h
@@ -2,7 +2,7 @@
 #include "Linear.h"
 #include "Activations.h"
 
-template<typename T, size_t rows, size_t hidden, size_t cols>
+template<typename T, int rows, int hidden, int cols>
 void ff(
 	T input[rows][cols],
 	T weights1[cols][hidden],
