+++ Layers/MatAdd.h
@@ -1,11 +1,11 @@
 
 #pragma once 
-template<typename T, size_t rows, size_t cols>
+template<typename T, int rows, int cols>
 void matadd(T A[rows][cols], T B[rows][cols], T result[rows][cols]) {
 matadd_outer_loop:
-	for (size_t i = 0; i < rows; i++) {
+	for (int i = 0; i < rows; i++) {
 	matadd_inner_loop:
-		for (size_t j = 0; j < cols; j++) {
+		for (int j = 0; j < cols; j++) {
 			result[i][j] = A[i][j] + B[i][j];
 		}
 	}
