+++ Tests/TestAttHead.h
@@ -2,7 +2,7 @@
 #include "TestUtils.h"
 #include "AttHead.h"
 
-template<typename T, size_t sequence_length, size_t token_length, size_t head_token_length>
+template<typename T, int sequence_length, int token_length, int head_token_length>
 void test_attHead(
 	std::string* matIn_filename,
 	std::string* weights_filename,
