+++ Tests/TestEncoder.h
@@ -2,11 +2,11 @@
 #include "TestUtils.h"
 #include "Encoder.h"
 
-template<typename T, size_t num_heads,
-	size_t sequence_length,
-	size_t token_length,
-	size_t head_token_length,
-	size_t hidden>
+template<typename T, int num_heads,
+	int sequence_length,
+	int token_length,
+	int head_token_length,
+	int hidden>
 	void test_encoder( 
 		std::string* matIn_filename, 
 		std::string* matMask_filename,
