+++ Tests/TestUtils.h
@@ -4,10 +4,10 @@
 #include <string>
 #include <sstream>
 
-template<typename T, size_t size>
+template<typename T, int size>
 void load_arr(T arr[size], std::string* filename) {
 	std::ifstream file(*filename);
-	for (size_t i = 0; i < size; i++) {
+	for (int i = 0; i < size; i++) {
 		std::string line;
 		std::getline(file, line);
 		std::stringstream stream;
@@ -17,14 +17,14 @@ void load_arr(T arr[size], std::string* filename) {
 }
 
 
-template<typename T, size_t size>
+template<typename T, int size>
 void compare_vec(T vec[size], std::string* vec_filename, std::string* log_filename) {
 	std::ifstream file(*vec_filename);
 	std::ofstream log(*log_filename);
 	int mismatch_count = 0;
 	double avg_error = 0;
 	bool good_result = true;
-	for (size_t i = 0; i < size; i++) {
+	for (int i = 0; i < size; i++) {
 		std::string line;
 		std::getline(file, line);
 		std::stringstream stream;
@@ -53,7 +53,7 @@ void compare_vec(T vec[size], std::string* vec_filename, std::string* log_filena
 	}
 }
 
-template<typename T, size_t rows, size_t cols>
+template<typename T, int rows, int cols>
 void compare_mat(T mat[rows][cols], std::string* mat_filename, std::string* log_filename) {
 	compare_vec<T, rows*cols>((T*) mat, mat_filename, log_filename);
 }
